Update mocked-backoffice skill with new env vars and patterns
- Update SKILL.md with VITE_EXAMPLE_PATH and VITE_UMBRACO_USE_MSW=on
- Update external-msw-handlers.md with createHandlers() pattern
- Update mock-repository-pattern.md with generic file names
- Remove outdated template folder (tree-example serves as working example)
- Fix tree-example to use new env var patterns consistently
