+++ plugins/umbraco-testing-skills/skills/umbraco-mocked-backoffice/patterns/external-msw-handlers.md
@@ -1,6 +1,6 @@
 # External MSW Handlers
 
-Add MSW handlers in a separate `mocks/` folder and load them via `VITE_EXTERNAL_MOCKS`. Your production code stays unchanged.
+Add MSW handlers in `src/msw/handlers.ts` and load them from your entry point when `VITE_UMBRACO_USE_MSW=on`. Your production code stays unchanged.
 
 > **Note**: For extensions using hey-api/OpenAPI clients, consider the [Mock Repository Pattern](./mock-repository-pattern.md) instead - it's simpler and avoids cross-origin issues.
 
@@ -8,24 +8,29 @@ Add MSW handlers in a separate `mocks/` folder and load them via `VITE_EXTERNAL_
 
 1. **Cross-Origin Requests**: MSW can only intercept same-origin requests. If your API client uses an absolute `baseUrl` (e.g., `https://localhost:44348`), you must configure it to use relative URLs in mock mode.
 
-2. **TypeScript Source Loading**: When loading TypeScript source directly via `VITE_EXTERNAL_EXTENSION`, dynamic imports with `.js` extensions may not resolve. This affects extensions with lazy-loaded repositories.
+2. **TypeScript Source Loading**: When loading TypeScript source directly via `VITE_EXAMPLE_PATH`, dynamic imports with `.js` extensions may not resolve. This affects extensions with lazy-loaded repositories.
 
 3. **hey-api Client**: The generated client uses CommonJS files that need an ESM wrapper for Vite.
 
 ## Directory Structure
 
 ```
-my-extension/
-├── src/                  # Production code
-│   ├── index.ts          # Exports manifests
+my-extension/Client/
+├── src/
+│   ├── index.ts          # Entry point (loads handlers when MSW enabled)
+│   ├── manifests.ts      # Your extension manifests
 │   ├── api/
 │   │   ├── client.gen.ts # hey-api generated client
 │   │   └── client/
 │   │       ├── index.cjs # CommonJS (generated)
 │   │       └── index.js  # ESM wrapper (you create this)
+│   ├── msw/              # MSW handlers
+│   │   └── handlers.ts   # Exports createHandlers() function
 │   └── ...
-└── mocks/                # MSW handlers for custom APIs
-    └── handlers.ts       # Must export `handlers` array
+└── tests/
+    └── msw/
+        ├── playwright.config.ts
+        └── my-extension.spec.ts
 ```
 
 ## Step 1: Create ESM Wrapper for hey-api Client
@@ -87,35 +92,50 @@ export const onInit: UmbEntryPointOnInit = (_host) => {
 
 ## Step 3: Create MSW Handlers
 
-Create `mocks/handlers.ts`:
+Create `src/msw/handlers.ts`:
 
 ```typescript
-// Get http and HttpResponse from window.MockServiceWorker (MSW v2, set by mocked backoffice)
-const { http, HttpResponse } = (window as any).MockServiceWorker || {};
+// Handler factory - gets http/HttpResponse from Umbraco's MSW instance at runtime
+export function createHandlers() {
+  const { http, HttpResponse } = (window as any).MockServiceWorker;
 
-if (!http) {
-  console.error('MSW not available');
+  // Mock data
+  const items = [
+    { id: 'item-1', name: '[MSW] Item 1', icon: 'icon-document' },
+    { id: 'item-2', name: '[MSW] Item 2', icon: 'icon-folder' },
+  ];
+
+  // Use relative paths (same-origin)
+  const API_PATH = '/umbraco/myextension/api/v1';
+
+  return [
+    http.get(`${API_PATH}/items`, () => {
+      return HttpResponse.json({ total: items.length, items });
+    }),
+
+    http.post(`${API_PATH}/items`, async ({ request }: { request: Request }) => {
+      const body = await request.json();
+      return HttpResponse.json({ id: 'new-id', ...body }, { status: 201 });
+    }),
+  ];
+}
+```
+
+## Step 4: Load Handlers from Entry Point
+
+Update `src/index.ts` to register handlers when MSW is enabled:
+
+```typescript
+// Register MSW handlers when running in MSW mode
+if (import.meta.env.VITE_UMBRACO_USE_MSW === 'on') {
+  import('./msw/handlers.js').then(({ createHandlers }) => {
+    const { addMockHandlers } = (window as any).MockServiceWorker;
+    addMockHandlers(...createHandlers());
+  });
 }
 
-// Mock data
-const items = [
-  { id: 'item-1', name: 'Item 1', icon: 'icon-document' },
-  { id: 'item-2', name: 'Item 2', icon: 'icon-folder' },
-];
-
-// Use relative paths (same-origin)
-const API_PATH = '/umbraco/myextension/api/v1';
-
-export const handlers = http ? [
-  http.get(`${API_PATH}/items`, () => {
-    return HttpResponse.json({ total: items.length, items });
-  }),
-
-  http.post(`${API_PATH}/items`, async ({ request }: { request: Request }) => {
-    const body = await request.json();
-    return HttpResponse.json({ id: 'new-id', ...body }, { status: 201 });
-  }),
-] : [];
+// Export manifests
+export const manifests = (await import('./manifests.js')).manifests;
 ```
 
 ## Running
@@ -123,62 +143,38 @@ export const handlers = http ? [
 ```bash
 cd /path/to/Umbraco-CMS/src/Umbraco.Web.UI.Client
 
-# IMPORTANT: Point to DIRECTORIES, not files
-# - VITE_EXTERNAL_EXTENSION: Directory containing index.ts
-# - VITE_EXTERNAL_MOCKS: Directory containing handlers.ts
-VITE_EXTERNAL_EXTENSION=/path/to/my-extension \
-VITE_EXTERNAL_MOCKS=/path/to/my-extension/mocks \
-npm run dev:external
+# Point to the Client directory containing src/index.ts
+VITE_EXAMPLE_PATH=/path/to/my-extension/Client VITE_UMBRACO_USE_MSW=on npm run dev
 ```
 
-### Common Mistakes
+### Environment Variables
 
-| Wrong | Correct |
-|-------|---------|
-| `VITE_EXTERNAL_MOCKS=.../mocks/handlers.ts` | `VITE_EXTERNAL_MOCKS=.../mocks` |
-| `VITE_EXTERNAL_EXTENSION=.../vite.config.ts` | `VITE_EXTERNAL_EXTENSION=...` (directory) |
+| Variable | Value | Purpose |
+|----------|-------|---------|
+| `VITE_EXAMPLE_PATH` | `/path/to/extension/Client` | Path to extension directory |
+| `VITE_UMBRACO_USE_MSW` | `on` | Enable MSW for core Umbraco APIs and your handlers |
 
-You'll see in the console:
+You'll see in the browser console:
 ```
-📦 Loading external extension from: /path/to/my-extension/src
-🎭 Loading external MSW handlers from: /path/to/my-extension/mocks
-🎭 Registered 2 external MSW handler(s)
+[MSW] Custom handlers registered
 ```
 
 ## Working Example
 
 See **tree-example** in `umbraco-backoffice/examples/tree-example/Client/`:
-- `index.ts` - Entry point that exports manifests from mock/
-- `src/` - Production code with OpenAPI-generated client
-- `mock/` - Mock repository implementation
-- `mocks/handlers.ts` - MSW handlers for tree API endpoints
-- `tests/` - Playwright tests (MSW, mock repo, and real E2E)
 
-```bash
-# For MSW handlers approach:
-VITE_EXTERNAL_EXTENSION=/path/to/tree-example/Client \
-VITE_EXTERNAL_MOCKS=/path/to/tree-example/Client/mocks \
-npm run dev:external
-
-# For mock repository approach (no VITE_EXTERNAL_MOCKS needed):
-VITE_EXTERNAL_EXTENSION=/path/to/tree-example/Client \
-npm run dev:external
-```
+| Path | Description |
+|------|-------------|
+| `src/index.ts` | Entry point with MSW handler registration |
+| `src/msw/handlers.ts` | MSW handlers using `createHandlers()` pattern |
+| `tests/msw/playwright.config.ts` | Playwright config with webServer |
+| `tests/msw/tree.spec.ts` | Test suite (6 tests) |
 
 ### Running the Tree Example Tests
 
 ```bash
 cd /path/to/tree-example/Client
+export UMBRACO_CLIENT_PATH=/path/to/Umbraco-CMS/src/Umbraco.Web.UI.Client
 
-# MSW tests (mocked backoffice must be running)
-npm run test:mocked:msw
-
-# Mock repository tests (mocked backoffice must be running)
-npm run test:mocked:repo
-
-# Real E2E tests (real Umbraco must be running)
-URL=https://localhost:44325 \
-UMBRACO_USER_LOGIN=admin@example.com \
-UMBRACO_USER_PASSWORD=yourpassword \
-npm run test:e2e
+npm run test:msw
 ```
