--- plugins/umbraco-testing-skills/skills/umbraco-mocked-backoffice/template/package.json
@@ -1,13 +1,0 @@
-{
-	"name": "umbraco-mocked-backoffice-template",
-	"version": "1.0.0",
-	"type": "module",
-	"description": "Example extension for testing with Umbraco mocked backoffice",
-	"scripts": {
-		"dev": "echo 'Run from Umbraco-CMS/src/Umbraco.Web.UI.Client with: VITE_EXTERNAL_EXTENSION=$(pwd)/src npm run dev:external'"
-	},
-	"devDependencies": {
-		"@umbraco-cms/backoffice": "^17.0.0",
-		"typescript": "~5.8.0"
-	}
-}
