--- plugins/umbraco-testing-skills/skills/umbraco-mocked-backoffice/template/tsconfig.json
@@ -1,16 +1,0 @@
-{
-	"compilerOptions": {
-		"target": "ES2022",
-		"module": "ESNext",
-		"moduleResolution": "bundler",
-		"lib": ["ES2022", "DOM", "DOM.Iterable"],
-		"strict": true,
-		"esModuleInterop": true,
-		"skipLibCheck": true,
-		"experimentalDecorators": true,
-		"useDefineForClassFields": false,
-		"declaration": false,
-		"noEmit": true
-	},
-	"include": ["src/**/*.ts"]
-}
